lambda/requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
boto3==1.34.0
Pillow==10.0.1
scikit-image==0.21.0 